ai-memory-hackathon-2026-02-07/pyproject.toml
[project]
name = "ai-memory-hackathon"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "cognee==0.4.1",
    "transformers",
    "qdrant-client",
    "cognee-community-vector-adapter-qdrant",
    "python-dotenv",
    "requests>=2.31",
    "boto3>=1.35",
]

[tool.uv]
package = false
